add balance sheet stats

+++ src/main/java/com/github/dusanzahoransky/stockanalyst/model/yahoo/balancesheet/AnnualLongTermDebt.java
@@ -21,7 +21,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 public class AnnualLongTermDebt {
 
     @JsonProperty("dataId")
-    private Integer dataId;
+    private Long dataId;
     @JsonProperty("asOfDate")
     private String asOfDate;
     @JsonProperty("periodType")
@@ -34,12 +34,12 @@ public class AnnualLongTermDebt {
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
     @JsonProperty("dataId")
-    public Integer getDataId() {
+    public Long getDataId() {
         return dataId;
     }
 
     @JsonProperty("dataId")
-    public void setDataId(Integer dataId) {
+    public void setDataId(Long dataId) {
         this.dataId = dataId;
     }
 
@@ -96,7 +96,7 @@ public class AnnualLongTermDebt {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(AnnualLongTermDebt.class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        sb.append(AnnualLongTermDebt.class.getName()).append('@').append(Long.toHexString(System.identityHashCode(this))).append('[');
         sb.append("dataId");
         sb.append('=');
         sb.append(((this.dataId == null)?"<null>":this.dataId));
