+++ src/main/java/com/github/dusanzahoransky/stockanalyst/model/yahoo/balancesheet/BalanceSheetStatement.java
@@ -89,7 +89,7 @@ public class BalanceSheetStatement {
     @JsonProperty("netReceivables")
     private NetReceivables netReceivables;
     @JsonProperty("maxAge")
-    private Integer maxAge;
+    private Long maxAge;
     @JsonProperty("longTermDebt")
     private LongTermDebt longTermDebt;
     @JsonProperty("inventory")
@@ -330,12 +330,12 @@ public class BalanceSheetStatement {
     }
 
     @JsonProperty("maxAge")
-    public Integer getMaxAge() {
+    public Long getMaxAge() {
         return maxAge;
     }
 
     @JsonProperty("maxAge")
-    public void setMaxAge(Integer maxAge) {
+    public void setMaxAge(Long maxAge) {
         this.maxAge = maxAge;
     }
 
@@ -382,7 +382,7 @@ public class BalanceSheetStatement {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(BalanceSheetStatement.class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        sb.append(BalanceSheetStatement.class.getName()).append('@').append(Long.toHexString(System.identityHashCode(this))).append('[');
         sb.append("intangibleAssets");
         sb.append('=');
         sb.append(((this.intangibleAssets == null)?"<null>":this.intangibleAssets));
