+++ src/main/java/com/github/dusanzahoransky/stockanalyst/model/yahoo/balancesheet/PreMarketChangePercent.java
@@ -57,7 +57,7 @@ public class PreMarketChangePercent {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(PreMarketChangePercent.class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        sb.append(PreMarketChangePercent.class.getName()).append('@').append(Long.toHexString(System.identityHashCode(this))).append('[');
         sb.append("raw");
         sb.append('=');
         sb.append(((this.raw == null)?"<null>":this.raw));
