+++ src/main/java/com/github/dusanzahoransky/stockanalyst/model/yahoo/balancesheet/QuoteType.java
@@ -174,7 +174,7 @@ public class QuoteType {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(QuoteType.class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        sb.append(QuoteType.class.getName()).append('@').append(Long.toHexString(System.identityHashCode(this))).append('[');
         sb.append("exchange");
         sb.append('=');
         sb.append(((this.exchange == null)?"<null>":this.exchange));
