+++ src/main/java/com/github/dusanzahoransky/stockanalyst/model/yahoo/balancesheet/RegularMarketPrice.java
@@ -18,19 +18,19 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 public class RegularMarketPrice {
 
     @JsonProperty("raw")
-    private Integer raw;
+    private Long raw;
     @JsonProperty("fmt")
     private String fmt;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
     @JsonProperty("raw")
-    public Integer getRaw() {
+    public Long getRaw() {
         return raw;
     }
 
     @JsonProperty("raw")
-    public void setRaw(Integer raw) {
+    public void setRaw(Long raw) {
         this.raw = raw;
     }
 
@@ -57,7 +57,7 @@ public class RegularMarketPrice {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(RegularMarketPrice.class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        sb.append(RegularMarketPrice.class.getName()).append('@').append(Long.toHexString(System.identityHashCode(this))).append('[');
         sb.append("raw");
         sb.append('=');
         sb.append(((this.raw == null)?"<null>":this.raw));
