+++ src/main/java/com/github/dusanzahoransky/stockanalyst/model/yahoo/balancesheet/TimeSeries.java
@@ -118,7 +118,7 @@ public class TimeSeries {
     @JsonProperty("annualCapitalStock")
     private List<AnnualCapitalStock> annualCapitalStock = new ArrayList<AnnualCapitalStock>();
     @JsonProperty("timestamp")
-    private List<Integer> timestamp = new ArrayList<Integer>();
+    private List<Long> timestamp = new ArrayList<Long>();
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
@@ -453,12 +453,12 @@ public class TimeSeries {
     }
 
     @JsonProperty("timestamp")
-    public List<Integer> getTimestamp() {
+    public List<Long> getTimestamp() {
         return timestamp;
     }
 
     @JsonProperty("timestamp")
-    public void setTimestamp(List<Integer> timestamp) {
+    public void setTimestamp(List<Long> timestamp) {
         this.timestamp = timestamp;
     }
 
@@ -475,7 +475,7 @@ public class TimeSeries {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(TimeSeries.class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        sb.append(TimeSeries.class.getName()).append('@').append(Long.toHexString(System.identityHashCode(this))).append('[');
         sb.append("annualTotalNonCurrentAssets");
         sb.append('=');
         sb.append(((this.annualTotalNonCurrentAssets == null)?"<null>":this.annualTotalNonCurrentAssets));
