+++ src/main/java/com/github/dusanzahoransky/stockanalyst/model/yahoo/balancesheet/YtdReturn.java
@@ -31,7 +31,7 @@ public class YtdReturn {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(YtdReturn.class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        sb.append(YtdReturn.class.getName()).append('@').append(Long.toHexString(System.identityHashCode(this))).append('[');
         sb.append("additionalProperties");
         sb.append('=');
         sb.append(((this.additionalProperties == null)?"<null>":this.additionalProperties));
